Crud,Auth,TokenGen,EntitiesRel

+++ Task-management-complete-crud/src/tasks/tasks.service.ts
@@ -20,13 +20,14 @@ export class TasksService {
   )  {}
 
   async  getTasks(
-    filterDto:GetTasksFilterDto
+    filterDto:GetTasksFilterDto,
+    user:User,
      
     ) : Promise <Task[]>{
     const {status,search}=filterDto;
     const query = this.taskRepository.createQueryBuilder('task');
 
-      
+      query.where('task.userId = :userId', {userId: user.id});
    
     
     if (status) {
@@ -43,24 +44,11 @@ export class TasksService {
    
   }
 
-/*
-    getAllTasks() {
-      return this.tasks;
-    }
-
-    getTasksWithFilters(filterDto: GetTasksFilterDto): Task[] {
-
-      const { status, search } = filterDto;
-
-      let tasks = this.getAllTasks;
-      
-      return this.tasks;
-
-    }
-*/
    
-async getTaskById(id: number): Promise <Task> {
-      const found = await this.taskRepository.findOne(id);
+async getTaskById(id: number,
+  user:User
+  ): Promise <Task> {
+      const found = await this.taskRepository.findOne({where : {id, userId: user.id}});
 
     if (!found) {
       throw new NotFoundException(`Task with ID "${id}" not founod`);
@@ -82,72 +70,31 @@ async getTaskById(id: number): Promise <Task> {
       task.user=user;
 
       await task.save();
-        delete task.user;
+      delete task.user;
       return task;
   }
 
-  async deleteTask(id:number) : Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id:number,
+    user:User,
+    ) : Promise<void> {
+    const result = await this.taskRepository.delete({id, userId: user.id});
     // console.log (result);
     if (result.affected=== 0) {
       throw new NotFoundException(`Task with ID "${id}" not founod`);
     }
     
   }
-    /*
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task>{
-      
-      return this.taskRepository.createTask(createTaskDto);
-    }
-    /*
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    const { title,description} = createTaskDto;
-      const task = new Task();
-      task.title = title;
-      task.description = description;
-      task.status = TaskStatus.OPEN;
-      await task.save();
-      return task;
-  }
-/*
-    getTaskById(id: string): Task {
-      const found = this.tasks.find(task => task.id === id);
-      if (!found){
-        throw new NotFoundException();
-      }
-      return found;
-    }
 
-
-    createTask(createTaskDto: CreateTaskDto): Task {
-      const { title, description } = createTaskDto;
-      const task: Task = {
-        id: uuidv4(),
-        title,
-        description,
-        status: TaskStatus.OPEN,
-      };
-      
-    this.tasks.push(task);
-    return task;
-    }
-    deleteTask(id: string): void {
-      const found =this.getTaskById(id);
-      this.tasks = this.tasks.filter(task => task.id !== id);
-    }*/
-
-   async  updateTaskStatus(id: number, status: TaskStatus): Promise <Task> {
-      const task = await this.getTaskById(id);
+   async  updateTaskStatus(
+     id: number, 
+     status: TaskStatus,
+     user: User,
+     ): Promise <Task> {
+      const task = await this.getTaskById(id,user);
       task.status = status;
        await task.save();
       return task;
-   }/*
-
-    updateTaskStatus(id: string, status: TaskStatus): Task {
-      const task = this.getTaskById(id);
-      task.status = status;
-      return task;
-    }*/
+   }
 
     
 }
